// src/app/cine-web/cine-svc/model/commentDTO.ts
/**
 * App API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: snapshot
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CommentDTO { 
    id?: number;
    userId?: number;
    username?: string;
    filmId?: number;
    commentText: string;
    createdAt?: string;
}